requirements.txt
alembic==1.12.0
aniso8601==9.0.1
attrs==23.1.0
Babel==2.13.0
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
email-validator==1.1.3
Flask==2.0.2
Flask-Babel==2.0.0
Flask-Login==0.5.0
Flask-Migrate==3.1.0
Flask-MySQLdb==2.0.0
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
greenlet==3.0.0
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
Mako==1.2.4
MarkupSafe==2.1.3
mysqlclient==2.2.0
python-decouple==3.5
pytz==2023.3.post1
referencing==0.30.2
rpds-py==0.10.6
six==1.16.0
SQLAlchemy==1.4.29
typing_extensions==4.8.0
Werkzeug==2.0.3
WTForms==3.0.1

# Google APIs (Gmail OAuth + API)
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0

# Token encryption
cryptography>=41.0.5

# OCR
pytesseract>=0.3.10
Pillow>=10.1.0

# OpenAI (LLM extraction + draft generation)
openai>=1.3.0

# Testing
pytest>=7.4.3
pytest-flask>=1.3.0
